% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssb_kommunedata.R
\name{get_ssb_dta}
\alias{get_ssb_dta}
\title{Hent data fra SSB}
\usage{
get_ssb_dta(ssb_tbl_id, aar, statistikkvariabel, kommuner, params = NULL)
}
\arguments{
\item{ssb_tbl_id}{Character. ID for tabellen som skal hentes fra SSB.}

\item{aar}{Character. Ett eller flere årstall som data skal hentes for.}

\item{statistikkvariabel}{Character. Kode for statistikkvariabelen som skal hentes.}

\item{kommuner}{Character vector. En vektor med kommunenummer (fire siffer).}

\item{params}{List. Valgfrie parametere for API-kallet.}
}
\value{
En data.frame med data fra SSB.
}
\description{
Denne funksjonen laster ned data fra Statistisk sentralbyrå (SSB) via deres API.
Funksjonen er laget for å hente data for kommuner.
}
