% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmn_inndeling.R
\name{kmn_inndeling}
\alias{kmn_inndeling}
\title{Get municipality classifications from SSB}
\usage{
kmn_inndeling(datoer)
}
\arguments{
\item{datoer}{String for the required date of the classification. Format must be "yyyy-mm-dd". For an inverval, provide two dates as a vector. If blank, will default to today's date}
}
\value{
A data frame with municipality name, number and year of validity
}
\description{
Get municipality classifications from SSB
}
\examples{
kommunestruktur_2020 <- kmn_inndeling(datoer = "2020-01-01")
head(kommunestruktur_2020)

}
