% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_doffin.R
\name{search_doffin}
\alias{search_doffin}
\title{Retrieve tender notices from Doffin using the Doffin API}
\usage{
search_doffin(..., api_key = NULL)
}
\arguments{
\item{...}{Additional parameters to be passed to the API.}

\item{api_key}{Optionally, supply the API key directly. If NULL, will look for the DOFFIN_API_KEY environment variable.}
}
\value{
A response object containing the search results.
}
\description{
Retrieve tender notices from Doffin using the Doffin API
}
\examples{
\dontrun{
if (Sys.getenv("DOFFIN_API_KEY") != "") {
  resp <- search_doffin()
}
}
}
