% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_doffin.R
\name{set_doffin_api_key}
\alias{set_doffin_api_key}
\title{Optionally prompts the user for the key if not supplied and askpass is available.}
\usage{
set_doffin_api_key(key = NULL)
}
\arguments{
\item{key}{The API key to set. If NULL, will prompt the user (if askpass is installed).}
}
\description{
Optionally prompts the user for the key if not supplied and askpass is available.
}
