% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmn_inndeling.R
\name{kmn_inndeling}
\alias{kmn_inndeling}
\title{Get municipality classifications from SSB}
\usage{
kmn_inndeling(aar)
}
\arguments{
\item{aar}{Integer vector with the years for which you want valid classifications.}
}
\value{
A data frame with municipality name, number and year of validity
}
\description{
Get municipality classifications from SSB
}
\examples{
kommunestruktur_2020 <- kmn_inndeling(aar = 2020)
head(kommunestruktur_2020)

}
