% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_doffin.R
\name{search_doffin}
\alias{search_doffin}
\title{Retrieve tender notices from Doffin using the Doffin API}
\usage{
search_doffin(...)
}
\arguments{
\item{...}{Additional parameters to be passed to the API.}
}
\value{
A response object containing the search results.
}
\description{
Retrieve tender notices from Doffin using the Doffin API
}
\examples{
resp <- search_doffin()
}
